+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.app/src/main/java/nts/uk/ctx/at/record/app/command/remainingnumber/empinfo/basicinfo/update/UpdateSpecialleave5informationCommand.java
@@ -1,5 +1,7 @@
 package nts.uk.ctx.at.record.app.command.remainingnumber.empinfo.basicinfo.update;
 
+import java.math.BigDecimal;
+
 import lombok.Getter;
 import nts.arc.time.GeneralDate;
 import nts.uk.shr.pereg.app.PeregEmployeeId;
@@ -15,15 +17,15 @@ public class UpdateSpecialleave5informationCommand {
 	
 	//	特別休暇管理
 	@PeregItem("IS00324")
-	private int useAtr;
+	private BigDecimal useAtr;
 	
 	//	付与設定
 	@PeregItem("IS00325")
-	private int appSet;
+	private BigDecimal appSet;
 	
 	//	付与日数
 	@PeregItem("IS00326")
-	private Integer grantDays;
+	private BigDecimal grantDays;
 	
 	//	付与テーブル
 	@PeregItem("IS00327")
