+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.app/src/main/java/nts/uk/ctx/at/record/app/command/remainingnumber/empinfo/basicinfo/update/UpdateSpecialleave7informationCommand.java
@@ -1,5 +1,7 @@
 package nts.uk.ctx.at.record.app.command.remainingnumber.empinfo.basicinfo.update;
 
+import java.math.BigDecimal;
+
 import lombok.Getter;
 import nts.arc.time.GeneralDate;
 import nts.uk.shr.pereg.app.PeregEmployeeId;
@@ -15,15 +17,15 @@ public class UpdateSpecialleave7informationCommand {
 	
 	//	特別休暇管理
 	@PeregItem("IS00338")
-	private int useAtr;
+	private BigDecimal useAtr;
 	
 	//	付与設定
 	@PeregItem("IS00339")
-	private int appSet;
+	private BigDecimal appSet;
 	
 	//	付与日数
 	@PeregItem("IS00340")
-	private Integer grantDays;
+	private BigDecimal grantDays;
 	
 	//	付与テーブル
 	@PeregItem("IS00341")
